+++ src/layout/navBars/breadcrumb/user.vue
@@ -78,7 +78,7 @@
 
 <script lang="ts">
 import { ref, getCurrentInstance, computed, reactive, toRefs, onMounted, defineComponent } from 'vue';
-import { useRouter } from 'vue-router';
+import { useRoute, useRouter } from 'vue-router';
 import { ElMessageBox, ElMessage } from 'element-plus';
 import screenfull from 'screenfull';
 import { useI18n } from 'vue-i18n';
@@ -97,6 +97,7 @@ export default defineComponent({
 	setup() {
 		const { t } = useI18n();
 		const { proxy } = <any>getCurrentInstance();
+		const route = useRoute();
 		const router = useRouter();
 		const stores = useUserInfo();
 		const storesThemeConfig = useThemeConfig();
@@ -165,7 +166,7 @@ export default defineComponent({
 						await resetRoute(); // 删除/重置路由
 						ElMessage.success(t('message.user.logOutSuccess'));
 						setTimeout(() => {
-							window.location.href = ''; // 去登录页
+							window.location.href = `#/login?redirect=${route.path}&params=${JSON.stringify(route.query ? route.query : route.params)}`; // 去登录页
 						}, 500);
 					})
 					.catch(() => {});